Changed var type from int to long

+++ Lecture1/calculator.c
@@ -2,15 +2,15 @@
 
 int main(void)
 {
-    int first_num, second_num;
+    long first_num, second_num;
     printf("Enter a value for x: ");
-    scanf("%d", &first_num);
-    printf("You entered, for x: %d \n", first_num);
+    scanf("%ld", &first_num);
+    printf("You entered, for x: %ld \n", first_num);
 
     printf("Enter a value for y: ");
-    scanf("%d", &second_num);
-    printf("You entered, for y: %d \n", second_num);
+    scanf("%ld", &second_num);
+    printf("You entered, for y: %ld \n", second_num);
 
-    printf("Sum of x and y is: %i \n", first_num+second_num);
+    printf("Sum of x and y is: %li \n", first_num+second_num);
     return 0;
 }
